Internal DNS work (#7104)
* re-orgnaize dns & allow setting of the nameserver group used for the internal DNS lookup

* prevent tests that potentially mutate or access shared state objectes from interfering with each other
