+++ Cargo.toml
@@ -378,6 +378,7 @@ serde_repr = "0.1"
 serde_with = "3.21.0"
 serde_yaml = "0.9.25"
 serde_plain = "1.0.2"
+serial_test = "3.5.0"
 sha2 = "0.10.3"
 sha3 = "0.10"
 si-scale = "0.3.0"
