+++ common/client-libs/gateway-client/src/client/websockets.rs
@@ -22,7 +22,7 @@ pub(crate) async fn connect_async_with_hickory(
 ) -> Result<(WebSocketStream<MaybeTlsStream<TcpStream>>, Response), GatewayClientError> {
     use tokio::net::TcpSocket;
 
-    let resolver = HickoryDnsResolver::default();
+    let resolver = HickoryDnsResolver::new();
     let uri =
         Url::parse(endpoint).map_err(|_| GatewayClientError::InvalidUrl(endpoint.to_owned()))?;
     let port: u16 = uri.port_or_known_default().unwrap_or(443);
