+++ common/http-api-client/Cargo.toml
@@ -50,6 +50,7 @@ nym-network-defaults = { workspace = true, optional = true }
 nym-http-api-client-macro = { workspace = true }
 
 [target."cfg(not(target_arch = \"wasm32\"))".dependencies]
+arc-swap = { workspace = true }
 h2 = { workspace = true }
 hickory-resolver = { workspace = true, features = ["https-ring", "tls-ring", "webpki-roots"] }
 
@@ -60,3 +61,4 @@ features = ["tokio"]
 
 [dev-dependencies]
 tracing-subscriber.workspace = true
+serial_test = { workspace = true }
